+++ src/components/drawers/EditKeyResult.vue
@@ -16,7 +16,7 @@
     </template>
 
     <template #page="{ pageIndex, prev }">
-      <form-section :hide-errors="true">
+      <form-section>
         <template v-if="pageIndex === 1">
           <form-component
             v-model="thisKeyResult.name"
@@ -79,7 +79,7 @@
           </div>
         </template>
 
-        <template v-if="!keyResult?.archived" #actions="{ handleSubmit }">
+        <template v-if="!keyResult?.archived" #actions="{ handleSubmit, submitDisabled }">
           <pkt-button
             v-if="pageIndex === 1"
             :text="$t('btn.cancel')"
@@ -97,7 +97,7 @@
 
           <btn-save
             :label="pageIndex === pageCount ? $t('btn.complete') : $t('btn.continue')"
-            :disabled="loading"
+            :disabled="submitDisabled || loading"
             variant="label-only"
             @click="handleSubmit(save)"
           />